feat: スキーマ堅牢化 + run_id自動生成
## P0: FK型整合
- match_results.id を BIGSERIAL に変更
- interaction_logs.match_result_id, feedback_events.match_result_id を BIGINT に統一
- Rust側の型も i64 に統一、SQLの配列キャストも ::bigint[] に変更

## match_results: 日次UPSERT化
- ON CONFLICT DO NOTHING → DO UPDATE に変更（同日は最新で上書き）
- updated_at, last_match_run_id カラム追加
- 「いつ・どの実行で更新されたか」を追跡可能に

## interaction_logs: 柔軟化
- UNIQUE制約を (talent_id, project_id, run_date) → (match_run_id, talent_id, project_id) に変更
- 同日複数実行でも全て記録可能（学習データ・監査ログの欠落防止）
- match_run_id を NOT NULL に（実行インスタンスIDとして必須）
- idx_interaction_logs_project_run_date インデックス追加

## run_date Generated Column
- match_results, interaction_logs で (created_at AT TIME ZONE 'Asia/Tokyo')::date として自動算出
- アプリ側で日付を渡す必要がなくなり、全バイナリで一貫した日付計算を保証

## run_id モジュール追加
- sr_common::run_id::get() でプロセス起動時のULIDを取得
- 手動でUUID生成する必要がなくなった
- ulid crateを依存に追加

## クエリ最適化
- fetch_match_results を run_date インデックスに最適化

## README更新
- ロードマップにスキーマ堅牢化を追加
- run_id使い方セクション追加
- データ設計の説明を拡充

+++ crates/sr-common/src/api/models/queue.rs
@@ -210,7 +210,7 @@ pub struct MatchResultRow {
 #[derive(Debug, Clone, Serialize)]
 pub struct InteractionLogRow {
     pub id: i64,
-    pub match_result_id: Option<i32>,
+    pub match_result_id: Option<i64>,
     pub talent_id: i64,
     pub project_id: i64,
     pub match_run_id: Option<String>,
@@ -227,7 +227,7 @@ pub struct InteractionLogRow {
 pub struct FeedbackEventRow {
     pub id: i64,
     pub interaction_id: Option<i64>,
-    pub match_result_id: Option<i32>,
+    pub match_result_id: Option<i64>,
     pub match_run_id: Option<String>,
     pub engine_version: Option<String>,
     pub config_version: Option<String>,
